docs: add notes to MAB template and update README
Renamed the incomplete sab.bt template to mab.bt as a more complete sab template was submitted. Added notes to the top of the mab template to explain this to other people referencing the repository. Updated the repository README to add Hellacopters as a contributor and explain the purpose of the main directories.